Web: Client: Appearance: HexColorPickerComponent was moved to shared/components as ColorPicker
